move surface-coordinates postprocessing functions from segmentation to multistack
