+++ bigfish/multistack/postprocess.py
@@ -11,11 +11,14 @@ import warnings
 
 import numpy as np
 import pandas as pd
+from scipy import ndimage as ndi
 
 import bigfish.stack as stack
 import bigfish.segmentation as segmentation
 
 from skimage.measure import regionprops
+from skimage.measure import find_contours
+from skimage.draw import polygon_perimeter
 
 
 # ### Object identification to sub-cellular regions ###
@@ -691,3 +694,377 @@ def summarize_extraction_results(fov_results, ndim, path_output=None):
         stack.save_data_to_csv(df, path_output)
 
     return df
+
+
+# Postprocessing
+
+def center_mask_coord(main, others=None):
+    """Center a 2-d binary mask (surface or boundaries) or a 2-d localization
+    coordinates array and pad it.
+
+    One mask or coordinates array should be at least provided (`main`). If
+    others masks or arrays are provided (`others`), they will be transformed
+    like `main`. All the provided masks should have the same shape.
+
+    Parameters
+    ----------
+    main : np.ndarray, np.uint or np.int or bool
+        Binary image with shape (y, x) or array of coordinates with shape
+        (nb_points, 2).
+    others : List(np.ndarray)
+        List of binary image with shape (y, x), array of coordinates with
+        shape (nb_points, 2) or array of coordinates with shape (nb_points, 3).
+
+    Returns
+    -------
+    main_centered : np.ndarray, np.uint or np.int or bool
+        Centered binary image with shape (y, x).
+    others_centered : List(np.ndarray)
+        List of centered binary image with shape (y, x), centered array of
+        coordinates with shape (nb_points, 2) or centered array of coordinates
+        with shape (nb_points, 3).
+
+    """
+    # TODO allow the case when coordinates do not represent external boundaries
+    # check parameters
+    stack.check_array(
+        main,
+        ndim=2,
+        dtype=[np.uint8, np.uint16, np.int64, bool])
+    stack.check_parameter(others=(list, type(None)))
+    if others is not None and len(others) != 0:
+        for x in others:
+            if x is None:
+                continue
+            stack.check_array(
+                x,
+                ndim=2,
+                dtype=[np.uint8, np.uint16, np.int64, bool])
+
+    # initialize parameter
+    marge = stack.get_margin_value()
+
+    # compute by how much we need to move the main object to center it
+    if main.shape[1] == 2:
+        # 'main' is already a 2-d coordinates array (probably locating the
+        # external boundaries of the main object)
+        main_coord = main.copy()
+    else:
+        # get external boundaries coordinates
+        main_coord = from_binary_to_coord(main)
+
+    # we get the minimum and maximum from external boundaries coordinates so
+    # we add 1 to the minimum and substract 1 to the maximum
+    min_y, max_y = main_coord[:, 0].min() + 1, main_coord[:, 0].max() - 1
+    min_x, max_x = main_coord[:, 1].min() + 1, main_coord[:, 1].max() - 1
+
+    # we compute the shape of the main object with a predefined marge
+    shape_y = max_y - min_y + 1
+    shape_x = max_x - min_x + 1
+    main_centered_shape = (shape_y + 2 * marge, shape_x + 2 * marge)
+
+    # center the main object
+    if main.shape[1] == 2:
+        # 'main' is a 2-d coordinates array
+        main_centered = main.copy()
+        main_centered[:, 0] = main_centered[:, 0] - min_y + marge
+        main_centered[:, 1] = main_centered[:, 1] - min_x + marge
+    else:
+        # 'main' is a 2-d binary matrix
+        main_centered = np.zeros(main_centered_shape, dtype=main.dtype)
+        crop = main[min_y:max_y + 1, min_x:max_x + 1]
+        main_centered[marge:shape_y + marge, marge:shape_x + marge] = crop
+
+    if others is None or len(others) == 0:
+        return main_centered, None
+
+    # center the others objects with the same transformation
+    others_centered = []
+    for other in others:
+        if other is None:
+            other_centered = None
+        elif other.shape[1] == 2:
+            # 'other' is a 2-d coordinates array
+            other_centered = other.copy()
+            other_centered[:, 0] = other_centered[:, 0] - min_y + marge
+            other_centered[:, 1] = other_centered[:, 1] - min_x + marge
+        elif other.shape[1] == 3 or other.shape[1] == 4:
+            # 'other' is a 3-d or 4-d coordinates
+            other_centered = other.copy()
+            other_centered[:, 1] = other_centered[:, 1] - min_y + marge
+            other_centered[:, 2] = other_centered[:, 2] - min_x + marge
+        else:
+            # 'other' is a 2-d binary matrix
+            other_centered = np.zeros(main_centered_shape, dtype=other.dtype)
+            crop = other[min_y:max_y + 1, min_x:max_x + 1]
+            other_centered[marge:shape_y + marge, marge:shape_x + marge] = crop
+        others_centered.append(other_centered)
+
+    return main_centered, others_centered
+
+
+def from_boundaries_to_surface(binary_boundaries):
+    """Fill in the binary matrix representing the boundaries of an object.
+
+    Parameters
+    ----------
+    binary_boundaries : np.ndarray, np.uint or np.int or bool
+        Binary image with shape (y, x).
+
+    Returns
+    -------
+    binary_surface : np.ndarray, bool
+        Binary image with shape (y, x).
+
+    """
+    # check parameters
+    stack.check_array(
+        binary_boundaries,
+        ndim=2,
+        dtype=[np.uint8, np.uint16, np.int64, bool])
+
+    # from binary boundaries to binary surface
+    binary_surface = ndi.binary_fill_holes(binary_boundaries)
+
+    return binary_surface
+
+
+def from_surface_to_boundaries(binary_surface):
+    """Convert the binary surface to binary boundaries.
+
+    Parameters
+    ----------
+    binary_surface : np.ndarray, np.uint or np.int or bool
+        Binary image with shape (y, x).
+
+    Returns
+    -------
+    binary_boundaries : np.ndarray, np.uint or np.int or bool
+        Binary image with shape (y, x).
+
+    """
+    # check parameters
+    stack.check_array(
+        binary_surface,
+        ndim=2,
+        dtype=[np.uint8, np.uint16, np.int64, bool])
+    original_dtype = binary_surface.dtype
+
+    # pad the binary surface in case object if on the edge
+    binary_surface_ = np.pad(binary_surface, [(1, 1)], mode="constant")
+
+    # compute distance map of the surface binary mask
+    distance_map = ndi.distance_transform_edt(binary_surface_)
+
+    # get binary boundaries
+    binary_boundaries_ = (distance_map < 2) & (distance_map > 0)
+    binary_boundaries_ = binary_boundaries_.astype(original_dtype)
+
+    # remove pad
+    binary_boundaries = binary_boundaries_[1:-1, 1:-1]
+
+    return binary_boundaries
+
+
+def from_binary_to_coord(binary):
+    """Extract coordinates from a 2-d binary matrix.
+
+    As the resulting coordinates represent the external boundaries of the
+    object, the coordinates values can be negative.
+
+    Parameters
+    ----------
+    binary : np.ndarray, np.uint or np.int or bool
+        Binary image with shape (y, x).
+
+    Returns
+    -------
+    coord : np.ndarray, np.int64
+        Array of boundaries coordinates with shape (nb_points, 2).
+
+    """
+    # check parameters
+    stack.check_array(
+        binary,
+        ndim=2,
+        dtype=[np.uint8, np.uint16, np.int64, bool])
+
+    # we enlarge the binary mask with one pixel to be sure the external
+    # boundaries of the object still fit within the frame
+    binary_ = np.pad(binary, [(1, 1)], mode="constant")
+
+    # get external boundaries coordinates
+    coord = find_contours(binary_, level=0)[0].astype(np.int64)
+
+    # remove the pad
+    coord -= 1
+
+    return coord
+
+
+def complete_coord_boundaries(coord):
+    """Complete a 2-d coordinates array, by generating/interpolating missing
+    points.
+
+    Parameters
+    ----------
+    coord : np.ndarray, np.int64
+        Array of coordinates to complete, with shape (nb_points, 2).
+
+    Returns
+    -------
+    coord_completed : np.ndarray, np.int64
+        Completed coordinates arrays, with shape (nb_points, 2).
+
+    """
+    # check parameters
+    stack.check_array(coord, ndim=2, dtype=[np.int64])
+
+    # for each array in the list, complete its coordinates using the scikit
+    # image method 'polygon_perimeter'
+    coord_y, coord_x = polygon_perimeter(coord[:, 0], coord[:, 1])
+    coord_y = coord_y[:, np.newaxis]
+    coord_x = coord_x[:, np.newaxis]
+    coord_completed = np.concatenate((coord_y, coord_x), axis=-1)
+
+    return coord_completed
+
+
+def from_coord_to_frame(coord, external_coord=True):
+    """Initialize a frame shape to represent coordinates values in 2-d matrix.
+
+    If coordinates represent the external boundaries of an object, we add 1 to
+    the minimum coordinate and substract 1 to the maximum coordinate in order
+    to build the frame. The frame centers the coordinates by default.
+
+    Parameters
+    ----------
+    coord : np.ndarray, np.int64
+        Array of cell boundaries coordinates with shape (nb_points, 2) or
+        (nb_points, 3).
+    external_coord : bool
+        Coordinates represent external boundaries of object.
+
+    Returns
+    -------
+    frame_shape : tuple
+        Shape of the 2-d matrix.
+    min_y : int
+        Value tu substract from the y coordinate axis.
+    min_x : int
+        Value tu substract from the x coordinate axis.
+    marge : int
+        Value to add to the coordinates.
+
+    """
+    # check parameter
+    stack.check_parameter(external_coord=bool)
+
+    # initialize marge
+    marge = stack.get_margin_value()
+
+    # from 2D coordinates boundaries to binary boundaries
+    if external_coord:
+        min_y, max_y = coord[:, 0].min() + 1, coord[:, 0].max() - 1
+        min_x, max_x = coord[:, 1].min() + 1, coord[:, 1].max() - 1
+    else:
+        min_y, max_y = coord[:, 0].min(), coord[:, 0].max()
+        min_x, max_x = coord[:, 1].min(), coord[:, 1].max()
+    shape_y = max_y - min_y + 1
+    shape_x = max_x - min_x + 1
+    frame_shape = (shape_y + 2 * marge, shape_x + 2 * marge)
+
+    return frame_shape, min_y, min_x, marge
+
+
+def from_coord_to_surface(cell_coord, nuc_coord=None, rna_coord=None,
+                          external_coord=True):
+    """Convert 2-d coordinates to a binary matrix with the surface of the
+    object.
+
+    If we manipulate the coordinates of the external boundaries, the relative
+    binary matrix has two extra pixels in each dimension. We compensate by
+    keeping only the inside pixels of the object surface.
+
+    If others coordinates are provided (nucleus and mRNAs), the relative
+    binary matrix is built with the same shape as the main coordinates (cell).
+
+    Parameters
+    ----------
+    cell_coord : np.ndarray, np.int64
+        Array of cell boundaries coordinates with shape (nb_points, 2).
+    nuc_coord : np.ndarray, np.int64
+        Array of nucleus boundaries coordinates with shape (nb_points, 2).
+    rna_coord : np.ndarray, np.int64
+        Array of mRNAs coordinates with shape (nb_points, 2) or
+        (nb_points, 3).
+    external_coord : bool
+        Coordinates represent external boundaries of object.
+
+    Returns
+    -------
+    cell_surface : np.ndarray, bool
+        Binary image of cell surface with shape (y, x).
+    nuc_surface : np.ndarray, bool
+        Binary image of nucleus surface with shape (y, x).
+    rna_binary : np.ndarray, bool
+        Binary image of mRNAs localizations with shape (y, x).
+    new_rna_coord : np.ndarray, np.int64
+        Array of mRNAs coordinates with shape (nb_points, 2) or (nb_points, 3).
+
+    """
+    # check parameters
+    stack.check_array(cell_coord, ndim=2, dtype=[np.int64])
+    if nuc_coord is not None:
+        stack.check_array(nuc_coord, ndim=2, dtype=[np.int64])
+    if rna_coord is not None:
+        stack.check_array(rna_coord, ndim=2, dtype=[np.int64])
+    stack.check_parameter(external_coord=bool)
+
+    # center coordinates
+    cell_coord_, [nuc_coord_, rna_coord_] = center_mask_coord(
+        main=cell_coord,
+        others=[nuc_coord, rna_coord])
+
+    # get the binary frame
+    frame_shape, min_y, min_x, marge = from_coord_to_frame(
+        coord=cell_coord_,
+        external_coord=external_coord)
+
+    # from coordinates to binary external boundaries
+    cell_boundaries_ext = np.zeros(frame_shape, dtype=bool)
+    cell_boundaries_ext[cell_coord_[:, 0], cell_coord_[:, 1]] = True
+    if nuc_coord_ is not None:
+        nuc_boundaries_ext = np.zeros(frame_shape, dtype=bool)
+        nuc_boundaries_ext[nuc_coord_[:, 0], nuc_coord_[:, 1]] = True
+    else:
+        nuc_boundaries_ext = None
+
+    # from binary external boundaries to binary external surface
+    cell_surface_ext = from_boundaries_to_surface(cell_boundaries_ext)
+    if nuc_boundaries_ext is not None:
+        nuc_surface_ext = from_boundaries_to_surface(nuc_boundaries_ext)
+    else:
+        nuc_surface_ext = None
+
+    # from binary external surface to binary surface
+    cell_surface = cell_surface_ext & (~cell_boundaries_ext)
+    if nuc_surface_ext is not None:
+        nuc_surface = nuc_surface_ext & (~nuc_boundaries_ext)
+    else:
+        nuc_surface = None
+
+    # center mRNAs coordinates
+    if rna_coord_ is not None:
+        rna_binary = np.zeros(frame_shape, dtype=bool)
+        if rna_coord_.shape[1] == 2:
+            rna_binary[rna_coord_[:, 0], rna_coord_[:, 1]] = True
+        else:
+            rna_binary[rna_coord_[:, 1], rna_coord_[:, 2]] = True
+        new_rna_coord = rna_coord_.copy()
+
+    else:
+        rna_binary = None
+        new_rna_coord = None
+
+    return cell_surface, nuc_surface, rna_binary, new_rna_coord
